+++ conf/evolutions/default/1.sql
@@ -3,7 +3,7 @@
 
 # --- !Ups
 
-create table "tempdatum" ("id" SERIAL PRIMARY KEY,"city" VARCHAR(254) NOT NULL,"min" DOUBLE PRECISION NOT NULL,"max" DOUBLE PRECISION NOT NULL,"date" DATE NOT NULL);
+create table "tempdatum" ("id" SERIAL PRIMARY KEY,"city" VARCHAR(254) NOT NULL,"min" INTEGER NOT NULL,"max" INTEGER NOT NULL,"date" DATE NOT NULL);
 
 # --- !Downs
